fix: unnecessary decimal value (#299)

+++ api/src/main/java/kr/toxicity/model/api/data/raw/ModelUV.java
@@ -28,7 +28,7 @@ import java.util.Objects;
  */
 public record ModelUV(
     @NotNull Float4 uv,
-    float rotation,
+    int rotation,
     @Nullable JsonElement texture
 ) {
